+++ lib/patternist/rspec.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+require 'rspec/core'
+require_relative 'rspec/shared_examples/controller'